requirements.txt
boto3
Pillow
opencv-python
watchdog